backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai 
pydantic
python-multipart
httpx
langchain
langchain-openai